# Allow toggling cache data compression at run time