--- a/crengine/include/lvtinydom.h
+++ b/crengine/include/lvtinydom.h
@@ -2400,4 +2400,7 @@ void runTinyDomUnitTests();
 /// pass true to enable CRC check for
 void enableCacheFileContentsValidation(bool enable);
 
+/// pass true to not compress data in cache file
+void doNotCompressCachedData(bool enable);
+
 #endif