See https://github.com/koreader/koreader/issues/3261#issuecomment-348710821 for why not compressing data saved in cache can be a good thing (speedup of many things, but uses more disk space).

It had to be specified at compile time with:
https://github.com/koreader/crengine/blob/6e37d80dd638211c35e7ed9892e476aaacc5a9a1/crengine/src/lvtinydom.cpp#L20-L23
but we will now be able to have the same effect as it being 0 by calling at run time, from cre.cpp: `compressCachedData(false)`.

Some updated numbers (with the same book as in https://github.com/koreader/koreader/issues/3261#issuecomment-348710821 - dunno why I get worse times today, may be I had a non-debug emulator build at the time...)

compression| cache before | load time | render time | close time | page change time| resulting cache size
---- | ---------------- | -------------------------- | ---------------- | ---------------- | ---------------- | ----------------
yes | no cache yet | 25 s | 47 s | 20 s | 1.5 s | 22 Mb
yes | existing cache | 2 s | 60 s | 1 s | 1.5 s | 22 Mb
no | no cache yet | 23 s | 28 s | 3 s | 1 s | 62 Mb
no | existing cache | 2 s | 28 s | 1 s | 1 s | 62 Mb

I did not really measure the "page change time", but one gets the feeling of it being really more responsive when no compression.
KOReader mem usage is nearly the same after startup and a few page changes at around 130M.

(I checked that, before this PR patch, and after, the above results for compression=yes are the same, so the added checks for `_compressCachedData` are negligeable).
As the magic headers are different, crengine deals itself with cleaning and rebuilding a cache file. Of course, we better not change it when a book is currently opened, but we can set it at start time.

Will be followed by this patch to cre.cpp:
```c
 static int initCache(lua_State *L) {
     const char *cache_path = luaL_checkstring(L, 1);
     int cache_size = luaL_checkint(L, 2);
+    bool compress_cached_data = true;
+    if (lua_isboolean(L, 3)) {
+        compress_cached_data = lua_toboolean(L, 3);
+    }
+
+    // Setting this to false uses more disk space for cache,
+    // but speed up rendering and page turns quite a bit
+    compressCachedData(compress_cached_data);

     ldomDocCache::init(lString16(cache_path), cache_size);

    return 0;
}
```
Then, we can simply use in credocument.lua:
```lua
function CreDocument:cacheInit()
    -- remove legacy cr3cache directory
    if lfs.attributes("./cr3cache", "mode") == "directory" then
        os.execute("rm -r ./cr3cache")
    end
-    cre.initCache(DataStorage:getDataDir() .. "/cache/cr3cache", 1024*1024*32)
+    cre.initCache(DataStorage:getDataDir() .. "/cache/cr3cache", 1024*1024*32, G_reader_settings:nilOrTrue("cre_compress_cached_data"))
end
```
So those interested can test how it does on our devices by adding to their settings.reader.lua:
`["cre_compress_cached_data"] = false,`