@poire-z 2018-02-09
Sorry, I'm not sure I get what you mean with "tacked on" - even after googling what "tack" means :)

May be you mean this new option is not integrated enough, and I should have gone further, like removing all the #if DOC_DATA_COMPRESSION_LEVEL so it feels less like a patch?
I can do that (I just wanted the patch to be small so it's easy to see there's no impact).