@Frenzie 2018-02-09
Literally it's using a tack/pin to stick someone on the wall or some other surface; metaphorically it's something that was appended later.

> May be you mean this new option is not integrated enough, and I should have gone further, like removing all the #if DOC_DATA_COMPRESSION_LEVEL so it feels less like a patch?

Yes and no. That is what I mean, but I'm not opposed to merging.