@poire-z 2018-02-09
Having a quick go with a hammer, and being more positive with `compressCachedData`.