@poire-z 2018-02-09
Cleaner indeed. Fixing PR top comment to reflect changes.