@poire-z 2018-02-10
Results on my Kobo Glo HD for the same book:

Load + render time with cache on SDCard / cache in RAM (/dev/ko_cache/):

compression| cache before | time on SDCard | time on RAM | resulting cache size
---- | ---------------- | -------------------------- | -------------------------- | ----------------
yes | no cache yet | 230 s | 229 s | 22 Mb
yes | existing cache | 167 s | 160 s | 22 Mb
no | no cache yet | 196 s | 190 s | 53 Mb
no | existing cache | 115 s | 115 s | 53 Mb

(not sure why the cache size is lower than on the emulator, 32bits vs 64bits arch may be)

For medium sized (2-3Mb) epub, the difference in opening time is not as much as I expected, and most often similar (10.5s vs 11s). Closing time (when going back to file manager) is better thus.
So, that's not as good as what I got on the emulator :| , and I guess we don't need to change the default (compressed cache).

Some other memory usage and timings:

Engine | process size | opening+render time
-------------------------- | -------------------------- | ----------------
crengine with cache | 72 Mb | see above
crengine with no cache | > 450 Mb | killed after 60s (cause not enough memory)
MuPDF 1st opening | 390 MB | 50 seconds
MuPDF 2nd opening in same process | > 440 MB | killed after 50s (cause not enough memory)