@poire-z 2018-03-01
For comparison, timings on the emulator before #100 (copied from first post):

compression| cache before | load time | render time | close time | page change time| resulting cache size
---- | ---------------- | -------------------------- | ---------------- | ---------------- | ---------------- | ----------------
yes | no cache yet | 25 s | 47 s | 20 s | 1.5 s | 22 Mb
yes | existing cache | 2 s | 60 s | 1 s | 1.5 s | 22 Mb
no | no cache yet | 23 s | 28 s | 3 s | 1 s | 62 Mb
no | existing cache | 2 s | 28 s | 1 s | 1 s | 62 Mb

After #100, with a `["cre_storage_size_factor"] = 6,` (needed for no WARNING with that book):

compression| cache before | load time | render time | close time | page change time| resulting cache size
---- | ---------------- | -------------------------- | ---------------- | ---------------- | ---------------- | ----------------
yes | no cache yet | 23 s | 20s | 3 s | 1 s | 22 Mb
yes | existing cache | 2 s | 1 s | 1 s | 1 s | 22 Mb
no | no cache yet | 23 s | 22 s | 2 s | 1 s | 61 Mb
no | existing cache | 2 s | 1 s | 1 s | 1 s | 61 Mb

So, I guesss disabling compression with `["cre_compress_cached_data"] = false,` does not really bring much anymore, except a bigger cache file.